archive/app/backend/requirements.txt
Django>=3.2.0,<4.0.0
djangorestframework>=3.12.0,<4.0.0
psycopg2-binary>=2.9.0,<3.0.0
Pillow>=9.0.0
django-cors-headers>=4.0.0